Move logics in `listeningShards` to `shardpreftable`
