+++ shardpreftable.go
@@ -1,10 +1,102 @@
 package main
 
 import (
+	"fmt"
 	"sync"
 
 	peer "github.com/libp2p/go-libp2p-peer"
 )
+
+const byteSize = 8 // in bits
+
+type ListeningShards struct {
+	shardBits []byte
+}
+
+func NewListeningShards() *ListeningShards {
+	return &ListeningShards{
+		shardBits: make([]byte, (numShards/8)+1),
+	}
+}
+
+func shardIDToBitIndex(shardID ShardIDType) (byte, byte, error) {
+	if shardID >= numShards {
+		return 0, 0, fmt.Errorf("Wrong shardID %v", shardID)
+	}
+	byteIndex := byte(shardID / byteSize)
+	bitIndex := byte(shardID % byteSize)
+	return byteIndex, bitIndex, nil
+}
+
+func (ls *ListeningShards) unsetShard(shardID ShardIDType) error {
+	byteIndex, bitIndex, err := shardIDToBitIndex(shardID)
+	if err != nil {
+		return fmt.Errorf("")
+	}
+	if int(byteIndex) >= len(ls.shardBits) {
+		return fmt.Errorf(
+			"(byteIndex=%v) >= (len(shardBits)=%v)",
+			byteIndex,
+			len(ls.shardBits),
+		)
+	}
+	// log.Printf("shardID=%v, byteIndex=%v, bitIndex=%v", shardID, byteIndex, bitIndex)
+	ls.shardBits[byteIndex] &= (^(1 << bitIndex))
+	return nil
+}
+
+func (ls *ListeningShards) setShard(shardID ShardIDType) error {
+	byteIndex, bitIndex, err := shardIDToBitIndex(shardID)
+	if err != nil {
+		return fmt.Errorf("")
+	}
+	if int(byteIndex) >= len(ls.shardBits) {
+		return fmt.Errorf(
+			"(byteIndex=%v) >= (len(shardBits)=%v)",
+			byteIndex,
+			len(ls.shardBits),
+		)
+	}
+	// log.Printf("shardID=%v, byteIndex=%v, bitIndex=%v", shardID, byteIndex, bitIndex)
+	ls.shardBits[byteIndex] |= (1 << bitIndex)
+	return nil
+}
+
+func (ls *ListeningShards) isShardSet(shardID ShardIDType) bool {
+	byteIndex, bitIndex, err := shardIDToBitIndex(shardID)
+	if err != nil {
+		fmt.Errorf("")
+	}
+	index := ls.shardBits[byteIndex] & (1 << bitIndex)
+	return index != 0
+}
+
+func (ls *ListeningShards) getShards() []ShardIDType {
+	shards := []ShardIDType{}
+	for shardID := ShardIDType(0); shardID < numShards; shardID++ {
+		if ls.isShardSet(shardID) {
+			shards = append(shards, shardID)
+		}
+	}
+	return shards
+}
+
+func (ls *ListeningShards) fromSlice(shards []ShardIDType) *ListeningShards {
+	for _, shardID := range shards {
+		ls.setShard(shardID)
+	}
+	return ls
+}
+
+// TODO: need checks against the format of bytes
+func (ls *ListeningShards) fromBytes(bytes []byte) *ListeningShards {
+	ls.shardBits = bytes
+	return ls
+}
+
+func (ls *ListeningShards) toBytes() []byte {
+	return ls.shardBits
+}
 
 // ShardPrefTable manages peers' shard preference
 type ShardPrefTable struct {
